+++ csrc/cuda/decode.h
@@ -27,12 +27,13 @@
 namespace retinanet {
 namespace cuda {
 
-int decode(int batchSize,
-    const void *const *inputs, void **outputs,
+int decode(int batch_size,
+    const void *const *inputs, void *const *outputs,
     size_t height, size_t width, size_t scale,
     size_t num_anchors, size_t num_classes,
     const std::vector<float> &anchors, float score_thresh, int top_n,
     void *workspace, size_t workspace_size, cudaStream_t stream);
+
 
 }
 }