+++ csrc/cuda/nms.cu
@@ -35,8 +35,8 @@
 #include <thrust/sequence.h>
 #include <thrust/execution_policy.h>
 #include <thrust/gather.h>
-#include <thrust/system/cuda/detail/cub/device/device_radix_sort.cuh>
-#include <thrust/system/cuda/detail/cub/iterator/counting_input_iterator.cuh>
+#include <cub/device/device_radix_sort.cuh>
+#include <cub/iterator/counting_input_iterator.cuh>
 
 namespace retinanet {
 namespace cuda {
@@ -80,9 +80,9 @@ __global__ void nms_kernel(
 }
 
 int nms(int batch_size,
-        const void *const *inputs, void **outputs,
-        size_t count, int detections_per_im, float nms_thresh,
-        void *workspace, size_t workspace_size, cudaStream_t stream) {
+  const void *const *inputs, void *const *outputs,
+  size_t count, int detections_per_im, float nms_thresh,
+  void *workspace, size_t workspace_size, cudaStream_t stream) {
 
   if (!workspace || !workspace_size) {
     // Return required scratch space size cub style
@@ -93,11 +93,11 @@ int nms(int batch_size,
     workspace_size += get_size_aligned<float>(count); // scores_sorted
   
     size_t temp_size_flag = 0;
-    thrust::cuda_cub::cub::DeviceSelect::Flagged((void *)nullptr, temp_size_flag,
-      thrust::cuda_cub::cub::CountingInputIterator<int>(count),
+    cub::DeviceSelect::Flagged((void *)nullptr, temp_size_flag,
+      cub::CountingInputIterator<int>(count),
       (bool *)nullptr, (int *)nullptr, (int *)nullptr, count);
     size_t temp_size_sort = 0;
-    thrust::cuda_cub::cub::DeviceRadixSort::SortPairsDescending((void *)nullptr, temp_size_sort,
+    cub::DeviceRadixSort::SortPairsDescending((void *)nullptr, temp_size_sort,
       (float *)nullptr, (float *)nullptr, (int *)nullptr, (int *)nullptr, count);
     workspace_size += std::max(temp_size_flag, temp_size_sort);
 
@@ -116,7 +116,7 @@ int nms(int batch_size,
     auto in_scores = static_cast<const float *>(inputs[0]) + batch * count;
     auto in_boxes = static_cast<const float4 *>(inputs[1]) + batch * count;
     auto in_classes = static_cast<const float *>(inputs[2]) + batch * count;
-
+    
     auto out_scores = static_cast<float *>(outputs[0]) + batch * detections_per_im;
     auto out_boxes = static_cast<float4 *>(outputs[1]) + batch * detections_per_im;
     auto out_classes = static_cast<float *>(outputs[2]) + batch * detections_per_im;
@@ -126,15 +126,14 @@ int nms(int batch_size,
       flags, thrust::placeholders::_1 > 0.0f);
 
     int *num_selected = reinterpret_cast<int *>(indices_sorted);
-    thrust::cuda_cub::cub::DeviceSelect::Flagged(workspace, workspace_size,
-      thrust::cuda_cub::cub::CountingInputIterator<int>(0),
+    cub::DeviceSelect::Flagged(workspace, workspace_size, cub::CountingInputIterator<int>(0),
       flags, indices, num_selected, count, stream);
     cudaStreamSynchronize(stream);
     int num_detections = *thrust::device_pointer_cast(num_selected);
 
     // Sort scores and corresponding indices
     thrust::gather(on_stream, indices, indices + num_detections, in_scores, scores);
-    thrust::cuda_cub::cub::DeviceRadixSort::SortPairsDescending(workspace, workspace_size,
+    cub::DeviceRadixSort::SortPairsDescending(workspace, workspace_size,
       scores, scores_sorted, indices, indices_sorted, num_detections, 0, sizeof(*scores)*8, stream);
 
     // Launch actual NMS kernel - 1 block with each thread handling n detections
@@ -144,7 +143,7 @@ int nms(int batch_size,
       indices_sorted, scores_sorted, in_classes, in_boxes);
 
     // Re-sort with updated scores
-    thrust::cuda_cub::cub::DeviceRadixSort::SortPairsDescending(workspace, workspace_size,
+    cub::DeviceRadixSort::SortPairsDescending(workspace, workspace_size,
       scores_sorted, scores, indices_sorted, indices, num_detections, 0, sizeof(*scores)*8, stream);
 
     // Gather filtered scores, boxes, classes