+++ extras/cppapi/export.cpp
@@ -3,10 +3,12 @@
 #include <fstream>
 #include <sstream>
 #include <vector>
+#include <glob.h>
 
 #include "../../csrc/engine.h"
 
 #define ROTATED false // Change to true for Rotated Bounding Box export
+#define COCO_PATH "/coco/coco2017/val2017" // Path to calibration images
 
 using namespace std;
 
@@ -40,8 +42,22 @@ Config parseArgs(int argc, char** argv) {
 	}
 	if (argc == 5) {
 		cfg.anchorsFile = argv[4];
-	}
+
 	return cfg;
+}
+
+
+inline vector<string> glob(int batch){
+	glob_t glob_result;
+	string path = string(COCO_PATH);
+	if(path.back()!='/') path+="/";
+	glob((path+"*").c_str(), (GLOB_TILDE | GLOB_NOSORT), NULL, &glob_result);
+	vector<string> calibration_files;
+	for(int i=0; i<batch; i++){
+		calibration_files.push_back(string(glob_result.gl_pathv[i]));
+	}
+	globfree(&glob_result);
+	return calibration_files;
 }
 
 
@@ -97,7 +113,8 @@ int main(int argc, char *argv[]) {
 	onnxFile.close();
 
 	// Define default RetinaNet parameters to use for TRT export
-	int batch = 1;
+	const vector<int> dynamic_batch_opts{1, 8, 16};
+	int calibration_batches = 2; // must be >= 1
 	float score_thresh = 0.05f;
 	int top_n = 1000;
 	size_t workspace_size =(1ULL << 30);
@@ -142,9 +159,8 @@ int main(int argc, char *argv[]) {
 		cout << endl;
 	}
 
-	// For now, assume we have already done calibration elsewhere 
-	// if we want to create an INT8 TensorRT engine, so no need 
-	// to provide calibration files or model name
+	// For INT8 calibration, after setting COCO_PATH on line 10:
+	// const vector<string> calibration_files = glob(calibration_batches*dynamic_batch_opts[1]);
 	const vector<string> calibration_files;
 	string model_name = "";
 
@@ -154,11 +170,12 @@ int main(int argc, char *argv[]) {
 		precision = "INT8";
 
 	cout << "Building engine..." << endl;
-	auto engine = retinanet::Engine(buffer, size, batch, precision, score_thresh, top_n,
+	auto engine = retinanet::Engine(buffer, size, dynamic_batch_opts, precision, score_thresh, top_n,
 		anchors, ROTATED, nms_thresh, detections_per_im, calibration_files, model_name, cfg.Int8CalibrationTable, verbose, workspace_size);
 	
 	cout << "Saving engine..." << endl;
 	engine.save(string(cfg.outputPlan));
+
 
 	delete [] buffer;