+++ retinanet/train.py
@@ -16,7 +16,8 @@ from .infer import infer
 def train(model, state, path, annotations, val_path, val_annotations, resize, max_size, jitter, batch_size, iterations,
           val_iterations, mixed_precision, lr, warmup, milestones, gamma, is_master=True, world=1, use_dali=True,
           verbose=True, metrics_url=None, logdir=None, rotate_augment=False, augment_brightness=0.0,
-          augment_contrast=0.0, augment_hue=0.0, augment_saturation=0.0, regularization_l2=0.0001, rotated_bbox=False):
+          augment_contrast=0.0, augment_hue=0.0, augment_saturation=0.0, regularization_l2=0.0001, rotated_bbox=False,
+          absolute_angle=False):
     'Train the model on the given dataset'
 
     # Prepare model
@@ -30,10 +31,12 @@ def train(model, state, path, annotations, val_path, val_annotations, resize, ma
     # Setup optimizer and schedule
     optimizer = SGD(model.parameters(), lr=lr, weight_decay=regularization_l2, momentum=0.9)
 
+    loss_scale = "dynamic" if use_dali else "128.0"
+
     model, optimizer = amp.initialize(model, optimizer,
                                       opt_level='O2' if mixed_precision else 'O0',
                                       keep_batchnorm_fp32=True,
-                                      loss_scale=128.0,
+                                      loss_scale=loss_scale,
                                       verbosity=is_master)
 
     if world > 1:
@@ -58,7 +61,7 @@ def train(model, state, path, annotations, val_path, val_annotations, resize, ma
                                             world, annotations, training=True, rotate_augment=rotate_augment,
                                             augment_brightness=augment_brightness,
                                             augment_contrast=augment_contrast, augment_hue=augment_hue,
-                                            augment_saturation=augment_saturation)
+                                            augment_saturation=augment_saturation, absolute_angle=absolute_angle)
     else:
         data_iterator = (DaliDataIterator if use_dali else DataIterator)(
             path, jitter, max_size, batch_size, stride,
@@ -75,10 +78,10 @@ def train(model, state, path, annotations, val_path, val_annotations, resize, ma
 
     # Create TensorBoard writer
     if logdir is not None:
-        from tensorboardX import SummaryWriter
+        from torch.utils.tensorboard import SummaryWriter
         if is_master and verbose:
             print('Writing TensorBoard logs to: {}'.format(logdir))
-        writer = SummaryWriter(logdir=logdir)
+        writer = SummaryWriter(log_dir=logdir)
 
     profiler = Profiler(['train', 'fw', 'bw'])
     iteration = state.get('iteration', 0)